fix: 엔티티 컬럼 타입, 오타 수정

+++ server/src/main/java/com/ssafy/floraserver/db/entity/Bookmark.java
@@ -16,8 +16,8 @@ public class Bookmark extends BaseEntity{
     private Long bId;
 
     // 연관관계 필요
-    private int uId;
+    private Long uId;
     
     // 연관관계 필요
-    private int sId;
+    private Long sId;
 }
